Simplified test
The test is both simpler and more reliable in catching problems. There's
no looping in the shader, so it's basically just the classic "message
passing" litmus test, just run in parallel and with a simple permutation
(multiplication by primes mod 64k) to mix up memory acccesses.

It also runs 1000 dispatches. This reliably catches thousands of
failures on my 5700XT.