Switch test to use atomics only
The previous version of the test had a race between a nonatomic store
and a nonatomic load, because the load was not conditional on the atomic
flag. Adding an "if" guard makes the weak behavior observed on AMD
5700 XT go away. If those memory accesses are changed to atomics, then
it's no longer a race, so the observed reorderings seem to be actual
failures of the GPU relative to the Vulkan memory model.

A typical run has 64M measurements (64k per dispatch, 1000 iterations),
of which about 5000 have a flag = 1 but data = 0.