Checkpoint variable warp size
We can use either 16 or 32 for warp size. The former is faster (on M1 Max). 8 is also a possibility but then the size of the histogram array would exceed the workgroup, so threads would need to deal with multiple histogram values.

Quick experiments with ELEMENTS_PER_THREAD show no gains for values other than 4.